:art: Move song code into song package
